+++ packages/graph-editor/src/components/flow/nodes/index.ts
@@ -13,6 +13,7 @@ import ContrastingNode from './color/contrastingNode.tsx';
 import CosNode from './math/cosNode.tsx';
 import CountNode from './math/countNode.tsx';
 import CreateColorNode from './color/createColorNode.tsx';
+import SetColorValueNode from './color/setColorValueNode.tsx';
 import DivisionNode from './math/divNode.tsx';
 import EnumeratedInputNode from './input/enumeratedInputNode.tsx';
 import HarmonicNode from './series/harmonicNode.tsx';
@@ -79,6 +80,8 @@ import ContrastingFromSetNode from './color/contrastingFromSetNode.tsx';
 import colorNameNode from './color/colorNameNode.tsx';
 import NearestTokensNode from './color/nearestNode.tsx';
 import ArrayPassUnit from './array/passUnit.tsx';
+import cssFunctions from './css/functionNode.tsx';
+import NameArrayNode from './array/name.tsx';
 
 export type NodeTypeLookup = Record<string, React.ReactNode | React.FC>;
 export type StateInitializer = Record<string, Record<string, any>>;
@@ -176,6 +179,7 @@ export const {
   blendNode,
   modNode,
   CreateColorNode,
+  SetColorValueNode,
   ScaleNode,
   EnumeratedInputNode,
   CountNode,
@@ -204,4 +208,6 @@ export const {
   NearestTokensNode,
   stringifyNode,
   ArrayPassUnit,
+  cssFunctions,
+  NameArrayNode,
 ]);