+++ packages/ui/src/redux/selectors/index.ts
@@ -5,6 +5,14 @@ import { createSelector } from 'reselect';
 export const theme = createSelector(ui, (state) => state.theme);
 export const tabs = (state: RootState) => state.ui.tabs;
 export const currentTab = createSelector(ui, (state) => state.currentTab);
+export const previewCodeSelector = createSelector(
+  ui,
+  (state) => state.previewCode,
+);
+export const showNodesPanelSelector = createSelector(
+  ui,
+  (state) => state.showNodesPanel,
+);
 
 export const outputSelector = (state: RootState) => state.editorOutput;